requirements.txt
discord.py==2.3.2
PyYAML==6.0.2
python-dotenv==1.0.1
